TST: Add more scalar tests
